Refactor Python modules into modular packages for better maintainability
Split large files (>300 lines) into smaller, focused modules:

**src/mycsp.py (326 lines) -> src/transforms/ package:**
- transforms/csp.py: MyCSP class (~130 lines)
- transforms/pca.py: MyPCA class (~70 lines)
- mycsp.py: Compatibility layer for backwards imports

**src/train.py (440 lines) -> src/training/ package:**
- training/core.py: train_and_evaluate, train_with_holdout
- training/comparison.py: compare_pipelines
- training/persistence.py: save_model, load_model
- training/subject.py: train_subject
- train.py: Compatibility layer

**src/visualization.py (394 lines) -> src/visualization/ package:**
- visualization/_base.py: _finalize_plot helper
- visualization/cv_plots.py: plot_cv_scores, plot_cv_detailed
- visualization/metrics_plots.py: plot_confusion_matrix, plot_training_summary
- visualization/comparison_plots.py: plot_pipeline_comparison

**tests/ reorganization:**
- tests/features/: Split test_features.py by extractor type
- tests/transforms/: Split test_mycsp.py (CSP/PCA)
- tests/pipeline/: Split test_pipeline.py by pipeline category

All 94 tests pass. Backwards compatibility maintained via re-exports.
