+++ src/train.py
@@ -1,424 +1,35 @@
 """
 Training module for EEG classification.
 
-Handles:
-- Model training with cross-validation
-- Hyperparameter selection
-- Model saving
+This module provides backwards compatibility. The actual implementations
+have been moved to the training package:
+    - training.core: train_and_evaluate, train_with_holdout
+    - training.comparison: compare_pipelines
+    - training.persistence: save_model, load_model
+    - training.subject: train_subject
+
+For new code, prefer importing directly from the training package:
+    from training import train_and_evaluate, train_subject
 """
 
-import os
-import logging
-from typing import Tuple, Dict, Optional, List, Any
-import joblib
-import numpy as np
-from numpy.typing import NDArray
-from sklearn.model_selection import cross_val_score, StratifiedKFold, train_test_split
-from sklearn.metrics import accuracy_score, classification_report, confusion_matrix
-from sklearn.pipeline import Pipeline
+# Re-export from training package for backwards compatibility
+from training import (
+    train_and_evaluate,
+    train_with_holdout,
+    compare_pipelines,
+    save_model,
+    load_model,
+    train_subject,
+)
 
-from constants import RANDOM_STATE, TARGET_ACCURACY
-from preprocess import preprocess_subject, get_run_type
-from pipeline import get_pipeline, list_pipelines
-from visualization import (plot_cv_scores, plot_confusion_matrix,
-                           plot_pipeline_comparison, plot_cv_detailed,
-                           plot_training_summary)
-
-# Configure logging
-logger = logging.getLogger(__name__)
-
-
-def train_and_evaluate(X: NDArray[np.float64], y: NDArray[np.int64],
-                       pipeline_name: str = 'csp_lda',
-                       cv: int = 5,
-                       verbose: bool = True,
-                       plot: bool = True,
-                       save_plots: bool = False,
-                       **pipeline_kwargs: Any) -> Tuple[Pipeline, NDArray[np.float64]]:
-    """
-    Train a pipeline with cross-validation.
-
-    Parameters
-    ----------
-    X : np.ndarray
-        EEG data of shape (n_epochs, n_channels, n_times)
-    y : np.ndarray
-        Labels of shape (n_epochs,)
-    pipeline_name : str
-        Name of the pipeline to use
-    cv : int
-        Number of cross-validation folds
-    verbose : bool
-        Whether to print results
-    plot : bool
-        Whether to plot results
-    save_plots : bool
-        Whether to save plots to disk
-    **pipeline_kwargs : dict
-        Additional arguments for the pipeline
-
-    Returns
-    -------
-    pipeline : Pipeline
-        Fitted pipeline (on all data)
-    scores : np.ndarray
-        Cross-validation scores
-    """
-    # Build pipeline
-    pipeline = get_pipeline(pipeline_name, **pipeline_kwargs)
-
-    # Cross-validation
-    cv_splitter = StratifiedKFold(n_splits=cv, shuffle=True, random_state=RANDOM_STATE)
-    scores = cross_val_score(pipeline, X, y, cv=cv_splitter, scoring='accuracy')
-
-    if verbose:
-        sep = '=' * 50
-        print(f"\n{sep}")
-        print(f"Pipeline: {pipeline_name}")
-        print(f"{sep}")
-        print(f"Cross-validation scores: {scores}")
-        std_dev = scores.std() * 2
-        print(f"Mean accuracy: {scores.mean():.4f} (+/- {std_dev:.4f})")
-        print(f"Min: {scores.min():.4f}, Max: {scores.max():.4f}")
-
-        # Check if target achieved
-        if scores.mean() >= TARGET_ACCURACY:
-            print(f"Target accuracy ({TARGET_ACCURACY:.0%}) ACHIEVED")
-        else:
-            gap = TARGET_ACCURACY - scores.mean()
-            print(f"Target accuracy ({TARGET_ACCURACY:.0%}) NOT achieved "
-                  f"- need {gap:.4f} more")
-
-    # Fit on all data for final model
-    pipeline.fit(X, y)
-
-    # Plot results
-    if plot:
-        save_path = f"plots/cv_scores_{pipeline_name}.png" if save_plots else None
-        plot_cv_scores(scores, title=f"CV Scores: {pipeline_name}", save_path=save_path)
-
-    return pipeline, scores
-
-
-def train_with_holdout(X: NDArray[np.float64], y: NDArray[np.int64],
-                       pipeline_name: str = 'csp_lda',
-                       test_size: float = 0.2,
-                       cv: int = 5,
-                       verbose: bool = True,
-                       plot: bool = True,
-                       save_plots: bool = False,
-                       **pipeline_kwargs: Any) -> Tuple[Pipeline, NDArray[np.float64], float]:
-    """
-    Train a pipeline with holdout test set.
-
-    Parameters
-    ----------
-    X : np.ndarray
-        EEG data
-    y : np.ndarray
-        Labels
-    pipeline_name : str
-        Name of the pipeline
-    test_size : float
-        Proportion of data for test set
-    cv : int
-        Number of cross-validation folds on training set
-    verbose : bool
-        Whether to print results
-    plot : bool
-        Whether to plot results
-    save_plots : bool
-        Whether to save plots to disk
-    **pipeline_kwargs : dict
-        Additional arguments for the pipeline
-
-    Returns
-    -------
-    pipeline : Pipeline
-        Fitted pipeline
-    cv_scores : np.ndarray
-        Cross-validation scores on training set
-    test_accuracy : float
-        Accuracy on holdout test set
-    """
-    # Split into train and test
-    X_train, X_test, y_train, y_test = train_test_split(
-        X, y, test_size=test_size, random_state=RANDOM_STATE, stratify=y
-    )
-
-    if verbose:
-        print("\nData split:")
-        print(f"  Training: {len(y_train)} epochs")
-        print(f"  Test: {len(y_test)} epochs")
-
-    # Build pipeline
-    pipeline = get_pipeline(pipeline_name, **pipeline_kwargs)
-
-    # Cross-validation on training set
-    cv_splitter = StratifiedKFold(n_splits=cv, shuffle=True, random_state=RANDOM_STATE)
-    cv_scores = cross_val_score(pipeline, X_train, y_train, cv=cv_splitter, scoring='accuracy')
-
-    if verbose:
-        print("\nCross-validation on training set:")
-        print(f"  Scores: {cv_scores}")
-        std_dev = cv_scores.std() * 2
-        print(f"  Mean: {cv_scores.mean():.4f} (+/- {std_dev:.4f})")
-
-    # Fit on all training data
-    pipeline.fit(X_train, y_train)
-
-    # Evaluate on test set
-    y_pred = pipeline.predict(X_test)
-    test_accuracy = accuracy_score(y_test, y_pred)
-
-    if verbose:
-        print("\nTest set evaluation:")
-        print(f"  Accuracy: {test_accuracy:.4f}")
-        print("\nClassification report:")
-        print(classification_report(y_test, y_pred))
-        print("\nConfusion matrix:")
-        print(confusion_matrix(y_test, y_pred))
-
-        # Check if target achieved
-        if test_accuracy >= TARGET_ACCURACY:
-            msg = f"\nTarget accuracy ({TARGET_ACCURACY:.0%}) ACHIEVED on test set"
-            print(msg)
-        else:
-            msg = f"\nTarget accuracy ({TARGET_ACCURACY:.0%}) NOT achieved on test set"
-            print(msg)
-
-    # Plot results
-    if plot:
-        # Plot CV scores
-        cv_save = f"plots/cv_scores_{pipeline_name}_holdout.png" if save_plots else None
-        cv_title = f"CV Scores (Training Set): {pipeline_name}"
-        plot_cv_scores(cv_scores, title=cv_title, save_path=cv_save)
-
-        # Plot confusion matrix
-        cm_save = f"plots/confusion_matrix_{pipeline_name}.png" if save_plots else None
-        cm_title = f"Confusion Matrix (Test Set): {pipeline_name}"
-        plot_confusion_matrix(
-            y_test, y_pred,
-            title=cm_title,
-            save_path=cm_save)
-
-        # Plot comprehensive summary
-        summary_save = f"plots/training_summary_{pipeline_name}.png" if save_plots else None
-        plot_training_summary(
-            cv_scores, y_test, y_pred,
-            pipeline_name=pipeline_name,
-            save_path=summary_save)
-
-    return pipeline, cv_scores, test_accuracy
-
-
-def compare_pipelines(
-        X: NDArray[np.float64], y: NDArray[np.int64],
-        cv: int = 5, verbose: bool = True,
-        plot: bool = True, save_plots: bool = False
-) -> Dict[str, Optional[Dict[str, Any]]]:
-    """
-    Compare all available pipelines.
-
-    Parameters
-    ----------
-    X : np.ndarray
-        EEG data
-    y : np.ndarray
-        Labels
-    cv : int
-        Number of cross-validation folds
-    verbose : bool
-        Whether to print results
-    plot : bool
-        Whether to plot results
-    save_plots : bool
-        Whether to save plots to disk
-
-    Returns
-    -------
-    results : dict
-        Dictionary mapping pipeline names to their scores
-    """
-    results: Dict[str, Optional[Dict[str, Any]]] = {}
-
-    if verbose:
-        print("\n" + "=" * 60)
-        print("Comparing all pipelines")
-        print("=" * 60)
-
-    for name in list_pipelines():
-        try:
-            pipeline = get_pipeline(name)
-            cv_splitter = StratifiedKFold(n_splits=cv, shuffle=True, random_state=RANDOM_STATE)
-            scores = cross_val_score(pipeline, X, y, cv=cv_splitter, scoring='accuracy')
-            results[name] = {
-                'mean': scores.mean(),
-                'std': scores.std(),
-                'scores': scores
-            }
-
-            if verbose:
-                print(f"\n{name}:")
-                print(f"  Mean accuracy: {scores.mean():.4f} (+/- {scores.std() * 2:.4f})")
-
-        except (ValueError, RuntimeError, TypeError) as e:
-            logger.warning(f"Pipeline {name} failed: {e}")
-            if verbose:
-                print(f"\n{name}: FAILED - {e}")
-            results[name] = None
-
-    # Find best pipeline
-    valid_results = {k: v for k, v in results.items() if v is not None}
-    if valid_results:
-        best = max(valid_results.items(), key=lambda x: x[1]['mean'])
-        if verbose:
-            best_acc = best[1]['mean']
-            print(f"\nBest pipeline: {best[0]} with {best_acc:.4f} accuracy")
-
-    # Plot results
-    if plot and valid_results:
-        # Bar plot comparison
-        comp_save = "plots/pipeline_comparison.png" if save_plots else None
-        plot_pipeline_comparison(results, save_path=comp_save)
-
-        # Detailed box plot
-        detail_save = "plots/pipeline_comparison_detailed.png" if save_plots else None
-        plot_cv_detailed(results, save_path=detail_save)
-
-    return results
-
-
-def save_model(pipeline: Pipeline, path: str, metadata: Optional[Dict[str, Any]] = None) -> None:
-    """
-    Save trained model to disk.
-
-    Parameters
-    ----------
-    pipeline : Pipeline
-        Trained sklearn pipeline
-    path : str
-        Path to save the model
-    metadata : dict
-        Additional metadata to save with the model
-    """
-    model_data = {
-        'pipeline': pipeline,
-        'metadata': metadata or {}
-    }
-
-    # Create directory if needed (handle both absolute and relative paths)
-    dir_path = os.path.dirname(path) or '.'
-    os.makedirs(dir_path, exist_ok=True)
-
-    # Use joblib for safer and more efficient serialization
-    joblib.dump(model_data, path)
-
-    logger.info(f"Model saved to: {path}")
-    print(f"Model saved to: {path}")
-
-
-def load_model(path: str) -> Tuple[Pipeline, Dict[str, Any]]:
-    """
-    Load trained model from disk.
-
-    Parameters
-    ----------
-    path : str
-        Path to the saved model
-
-    Returns
-    -------
-    pipeline : Pipeline
-        Trained sklearn pipeline
-    metadata : dict
-        Metadata saved with the model
-    """
-    # Use joblib for loading
-    model_data = joblib.load(path)
-
-    return model_data['pipeline'], model_data['metadata']
-
-
-def train_subject(subject: int, runs: List[int],
-                  pipeline_name: str = 'csp_lda',
-                  model_dir: str = 'models',
-                  cv: int = 5,
-                  plot: bool = True,
-                  save_plots: bool = False,
-                  **pipeline_kwargs: Any) -> Tuple[Pipeline, NDArray[np.float64]]:
-    """
-    Complete training pipeline for a subject.
-
-    Parameters
-    ----------
-    subject : int
-        Subject number
-    runs : list
-        List of run numbers
-    pipeline_name : str
-        Name of the pipeline to use
-    model_dir : str
-        Directory to save the model
-    cv : int
-        Number of cross-validation folds
-    plot : bool
-        Whether to plot results
-    save_plots : bool
-        Whether to save plots to disk
-    **pipeline_kwargs : dict
-        Additional arguments for the pipeline
-
-    Returns
-    -------
-    pipeline : Pipeline
-        Trained pipeline
-    scores : np.ndarray
-        Cross-validation scores
-    """
-    sep = '=' * 60
-    print(f"\n{sep}")
-    print("Training BCI model")
-    print(f"Subject: {subject}, Runs: {runs}")
-    print(f"Pipeline: {pipeline_name}")
-    print(f"{sep}")
-
-    # Preprocess data
-    print("\nLoading and preprocessing data...")
-    X, y, epochs = preprocess_subject(subject, runs)
-
-    print(f"Data shape: {X.shape}")
-    print(f"Labels: {np.unique(y)}")
-    unique, counts = np.unique(y, return_counts=True)
-    class_dist = dict(zip(unique, counts))
-    print(f"Class distribution: {class_dist}")
-
-    # Train and evaluate
-    pipeline, scores = train_and_evaluate(
-        X, y, pipeline_name=pipeline_name, cv=cv,
-        plot=plot, save_plots=save_plots, **pipeline_kwargs
-    )
-
-    # Save model
-    run_type = get_run_type(runs[0])
-    model_path = os.path.join(model_dir, f"model_s{subject}_r{runs[0]}_{pipeline_name}.pkl")
-
-    metadata = {
-        'subject': subject,
-        'runs': runs,
-        'run_type': run_type,
-        'pipeline_name': pipeline_name,
-        'cv_scores': scores.tolist(),
-        'cv_mean': scores.mean(),
-        'n_epochs': len(y),
-        'n_channels': X.shape[1],
-        'n_times': X.shape[2],
-        'classes': np.unique(y).tolist()
-    }
-
-    save_model(pipeline, model_path, metadata)
-
-    return pipeline, scores
+__all__ = [
+    'train_and_evaluate',
+    'train_with_holdout',
+    'compare_pipelines',
+    'save_model',
+    'load_model',
+    'train_subject',
+]
 
 
 if __name__ == "__main__":
