+++ tests/features/__init__.py
@@ -1,0 +1,10 @@
+"""
+Tests for feature extraction modules.
+
+This package contains tests for:
+- PSDExtractor
+- BandPowerExtractor
+- LogVarianceExtractor
+- FlattenExtractor
+- sklearn compatibility
+"""
