+++ tests/pipeline/__init__.py
@@ -1,0 +1,9 @@
+"""
+Tests for pipeline modules.
+
+This package contains tests for:
+- Pipeline factory functions
+- CSP-based pipelines
+- Feature extraction pipelines
+- Integration tests
+"""
