+++ tests/transforms/__init__.py
@@ -1,0 +1,8 @@
+"""
+Tests for transforms modules.
+
+This package contains tests for:
+- MyCSP (Common Spatial Patterns)
+- MyPCA (Principal Component Analysis)
+- sklearn compatibility
+"""
